feat(curation): 패키지 변경 및 코드 변경
#141
