+++ src/main/java/com/thehyundai/thepet/domain/subscription/CurationService.java
@@ -1,5 +1,14 @@
 package com.thehyundai.thepet.domain.subscription;
 
+import java.util.List;
+
 public interface CurationService {
+
     CurationVO createCuration(CurationVO curationVO);
+
+    CurationVO showCurationOfCurrMonth();
+
+    List<CurationVO> showCurationOfLastOneYear();
+
+    CurationVO showCurationDetail(String curationId);
 }
